Fixed inaccuracy related to function construct
This "as long as it's in order" statement is incorrect. The counter example is `fn` which from the example is a valid construction to define a function. Note: `fn` are letters in `function` but not in order.

Therefore, this qualification should be removed (or the example should be removed if it is out of spec)